Move config.js into config/ dir
Add a src/annotator/config/ directory to contain all config-related code
for the src/annotator/ part of the code. Move config.js into the
directory.  We're going to be adding more config-related files in this
directory later.
